move history logic to history.py, keep __init__.py empty
